Added Roster Functionality and Custom Toolbar
added a new team roster activity with a recyclerview that will display
a scrollable list of a specific teams players along with their stats.

Current main activity and starting screen is set to the RosterActivity.

layed the groundwork for adding a functional toolbar that will allow for navigation between activities.